requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
APScheduler==3.10.4
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
astroid==2.15.6
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
bitarray==2.7.6
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
croniter==1.4.1
cronitor==4.6.0
cytoolz==0.12.2
dill==0.3.7
et-xmlfile==1.1.0
eth-abi==4.1.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.0
exceptiongroup==1.1.2
fastapi==0.100.0
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
hexbytes==0.3.1
httptools==0.6.0
humanize==4.7.0
idna==3.4
inflect==7.0.0
iniconfig==2.0.0
isort==5.12.0
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
lazy-object-proxy==1.9.0
lru-dict==1.2.0
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
numpy==1.25.1
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
parsimonious==0.9.0
platformdirs==3.10.0
pluggy==1.2.0
protobuf==4.23.4
psycopg2-binary==2.9.6
pycodestyle==2.11.0
pycryptodome==3.18.0
pydantic==2.1.1
pydantic-settings==2.0.2
pydantic_core==2.4.0
pyflakes==3.1.0
pylint==2.17.5
pytest==7.4.0
python-cron==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyunormalize==15.0.0
PyYAML==6.0.1
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.9.2
schedule==1.2.0
six==1.16.0
sniffio==1.3.0
sqlacodegen==2.3.0.post1
SQLAlchemy==1.4.49
starlette==0.27.0
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
typing_extensions==4.7.1
tzdata==2023.3
upstash-redis==0.15.0
urllib3==1.26.6
uvicorn==0.23.1
uvloop==0.17.0
watchfiles==0.19.0
web3==6.7.0
websockets==11.0.3
wrapt==1.15.0
yarl==1.9.2
